+++ src/budge/transfer.py
@@ -1,7 +1,7 @@
 from dataclasses import InitVar, dataclass, field
 
 from .account import Account
-from .transaction import RecurringTransaction, Transaction
+from .transaction import RepeatingTransaction, Transaction
 
 
 @dataclass(kw_only=True)
@@ -28,7 +28,7 @@ class Transfer(Transaction):
 
 
 @dataclass(kw_only=True)
-class RecurringTransfer(Transfer, RecurringTransaction):
+class RepeatingTransfer(Transfer, RepeatingTransaction):
     """
     A transfer between two accounts that repeats on a schedule described by a
     `dateutil.rrule.rrule`.
@@ -36,17 +36,17 @@ class RecurringTransfer(Transfer, RecurringTransaction):
 
     def __post_init__(self, from_account: Account, to_account: Account):
         """
-        Create the from and to recurring transactions, add them to the
-        respective accounts, and set their parent to this recurring transfer.
+        Create the from and to repeating transactions, add them to the
+        respective accounts, and set their parent to this repeating transfer.
         """
-        self.from_transaction = RecurringTransaction(
+        self.from_transaction = RepeatingTransaction(
             -self.amount, self.description, schedule=self.schedule
         )
-        self.to_transaction = RecurringTransaction(
+        self.to_transaction = RepeatingTransaction(
             self.amount, self.description, schedule=self.schedule
         )
 
         self.from_transaction.parent = self.to_transaction.parent = self
 
-        from_account.recurring_transactions.append(self.from_transaction)
-        to_account.recurring_transactions.append(self.to_transaction)
+        from_account.repeating_transactions.append(self.from_transaction)
+        to_account.repeating_transactions.append(self.to_transaction)